+++ ChangeLog.md
@@ -11,8 +11,10 @@
   article. ([#33](https://github.com/davep/oldnews/pull/33))
 - Added the `CopyHomePageToClipboard` command for copying the current
   subscription's homepage URL to the clipboard.
+  ([#43](https://github.com/davep/oldnews/pull/43))
 - Added the `CopyFeedToClipboard` command for copying the current
   subscription's feed URL to the clipboard.
+  ([#43](https://github.com/davep/oldnews/pull/43))
 
 ## v0.3.0